- fixed NPE while closing the dialog
git-svn-id: svn://helios.nb.team-ulm.de/uploadSystem@154 a202d7e1-1033-44dc-8c76-cd45ceb0298f

+++ trunk/src/com/teamulm/uploadsystem/client/gui/ViewGalleryDialog.java
@@ -156,11 +156,12 @@ public class ViewGalleryDialog extends Dialog {
 		}
 		this.tmpImages.clear();
 		for (File file : this.tmpFiles) {
-			file.delete();
+			if (null != file) {
+				file.delete();
+			}
 		}
 		this.tmpFiles.clear();
 		this.iconCross.dispose();
 		return super.close();
 	}
-
 }
